requirements.txt
# base tools
pip >= 21.3.1
wheel>=0.37.1
setuptools>=59.6.0

# Flask structure
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
python-dotenv==0.20.0

# Database
SQLAlchemy==1.4.54
pyodbc==4.0.39

# GraphQL
ariadne==0.16.0
graphql-core==3.2.6

# Test tools
pytest==7.0.1

# other tools
Ariadne==0.16.0
bcrypt==3.2.0
PyJWT==2.4.0
